Extract admin tests into own test files
